+++ templates/back/signalement/view/photos-documents.html.twig
@@ -51,7 +51,7 @@
                  data-user="{{ photo.user.nomComplet ?? 'N/R' }}" data-mail="{{ photo.createdAt is defined ? photo.createdAt|date('d.m.Y') : 'N/R' }}
 "
                  style="background: url('{{ asset('_up/'~photo.filename~'/'~signalement.uuid~'?variant=thumb') }}')no-repeat center center/cover">
-                <a href="{{ asset('_up/'~photo.filename~'/'~signalement.uuid) }}"
+                <a href="{{ asset('_up/'~photo.filename~'/'~signalement.uuid~'?variant=resize') }}"
                     class="fr-btn fr-btn--sm fr-icon-eye-line img-box" title="Voir la photo"
                     target="_blank" rel="noopener"></a>
                 {% if is_granted('ROLE_ADMIN_PARTNER') or (isAffected and isAccepted) %}
